Log helper warnings setup tests
Added tests to make sure the correct handlers are added to the warnings logger.
Skipped tests which check warnings are logged due to issues when running
with all tests in pytest, they run sucessfully when each test is ran
individually.
Removed "not serial" flag from tox as the serial tests are now skipped
instead.