review ready
using BaseConfig I think has made the overall solution more complex. It may be down to my lack of knowledge with BaseConfig but the test has become more complex. Everything passes bar the fact that this is required:
        text = _Metadata.load_yaml(meta_out)
        if hasattr(text, "dict"):
            # text = text.dict()
            text = text.model_dump(mode="json")
load_yaml returns an object so using either .dict or .model_dump is required. Both result in a mismatch between the URL provided by the "user". It's down to a trailing slash after the URL which is added when using .dict or .model_dump. The best solution would be to not use BaseConfig but I may be missing something with the package and the load_yaml methods.