[fix] BaseConfig convert to Path for file open
Previously switched to using `path.open` instead of built-in open function which will not work if path is a str.
The type annotation specifies it should be a Path object (not a string) but if people don't use MyPy they may not notice this and it's simple enough to attempt to convert the object to Path before calling open so it will work with strings even if we don't encourage it.
Added unit test which fails before this change is made but passes after.